refactor: extract get_stock_ratings into market/ratings.py (#575)
Moves get_stock_ratings out of robinhood_market_data_tools.py into a focused
market/ratings.py module. The original module re-exports the function for legacy
import compatibility. server/app.py imports directly from market.ratings. Tests
updated to import from market.ratings and patch market.ratings.* namespaces.
